pyproject.toml
[project]
name = "beesbotweb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "langchain>=0.3.26",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.26",
    "langchain-openai>=0.3.25",
    "langgraph>=0.4.8",
    "openai>=1.91.0",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
]
